+++ server/routes/autosave.js
@@ -1,11 +1,11 @@
 // Document autosaves
 // Server imports
-const { error } = require("console");
 const express = require("express");
 const router = express.Router();
 const fs = require("fs");
 const path = require("path");
 const serverMaster = require("../serverMaster");
+const logger = require("../logger");
 
 // Data paths
 const rootPath = path.join(__dirname, "../../");
@@ -13,14 +13,23 @@ const dataFolderPath = path.join(rootPath, "data");
 const autosavesFolderPath = path.join(dataFolderPath, "autosaves");
 const masterFilePath = path.join(dataFolderPath, "master.json");
 
-// Gets the masterfile and returns it.
+// Gets the master.json and returns it.
 async function getMasterFile() {
+  if (serverMaster.detailLogs) {
+    logger.info("Getting master.json");
+  }
   const rawMasterFile = await fs.readFileSync(masterFilePath, "utf-8");
   return JSON.parse(rawMasterFile);
 }
 
 // Adds unsaved filenames to the master.
 async function addUnsavedToMaster(filename) {
+  if (serverMaster.detailLogs) {
+    logger.info(
+      { Name: filename },
+      "Adding unsaved filename to master.json...",
+    );
+  }
   let masterFile = await getMasterFile();
   let unsavedFiles = masterFile[0].unsavedFiles; // Adds file to unsavedFiles array.
   let addFile = true;
@@ -43,16 +52,29 @@ async function addUnsavedToMaster(filename) {
         JSON.stringify(masterFile),
         "utf-8",
       );
-      console.log("Successfully added unsaved filename to master.json.");
+      if (serverMaster.successLogs) {
+        logger.info(
+          { Name: filename },
+          "Added unsaved filename to master.json.",
+        );
+      }
       return true;
     } catch (err) {
-      console.error("Failed to add unsaved filename to master.json.");
-      console.error(err);
+      logger.error(
+        { Name: filename, Error: err },
+        "Failed to add unsaved filename to master.json.",
+      );
       res.status(500).json({ error: err });
       return false;
     }
   } else {
     // Returns true if the file is already in the master.
+    if (serverMaster.detailLogs) {
+      logger.info(
+        { Name: filename },
+        "Filename is already in master.json. No changes made.",
+      );
+    }
     return true;
   }
 }
@@ -64,19 +86,36 @@ router.post("/api/autosave", async (req, res) => {
   const saveArray = [saveData];
   const autosaveFilePath = path.join(autosavesFolderPath, name); // Full path to the autosave file location.
 
+  if (serverMaster.detailLogs) {
+    logger.info(
+      { Name: name, Path: folderPath },
+      "Recieved autosave create request.",
+    );
+  }
+
   // If the file is successfully added to the unsavedFiles array, it creates the autosave.
   if (addUnsavedToMaster(name)) {
     try {
+      if (serverMaster.detailLogs) {
+        logger.info({ Name: name, Path: folderPath }, "Writing autosave...");
+      }
+
       await fs.writeFileSync(
         autosaveFilePath,
         JSON.stringify(saveArray),
         "utf-8",
       );
-      console.log("Successfully created autosave.");
+
+      if (serverMaster.successLogs) {
+        logger.info({ Name: name, Path: folderPath }, "Created autosave.");
+      }
+
       res.json({ success: true });
     } catch (err) {
-      console.error("Failed to create autosave.");
-      console.error(err);
+      logger.error(
+        { Name: name, Path: folderPath, Error: err },
+        "Failed to create autosave.",
+      );
       res.status(500).json({ error: err });
     }
   }
@@ -85,6 +124,10 @@ router.post("/api/autosave", async (req, res) => {
 // Removes autosave from the server.
 router.delete("/api/removeAutosave", async (req, res) => {
   const { name } = req.body;
+
+  if (serverMaster.detailLogs) {
+    logger.info({ Name: name }, "Recieved autosave delete request.");
+  }
 
   let masterFile = await getMasterFile();
   let unsavedFiles = masterFile[0].unsavedFiles;
@@ -99,15 +142,29 @@ router.delete("/api/removeAutosave", async (req, res) => {
   masterFile[0].unsavedFiles = unsavedFiles;
 
   try {
+    if (serverMaster.detailLogs) {
+      logger.info(
+        { Name: name },
+        "Removing saved filename from master.json...",
+      );
+    }
+
     // Updates master.json on the fs.
     await fs.writeFileSync(masterFilePath, JSON.stringify(masterFile), "utf-8");
-    console.log("Successfully removed filename from master.json.");
+
+    if (serverMaster.detailLogs) {
+      logger.info({ Name: name }, "Removed saved filename from master.json.");
+      logger.info({ Name: name }, "Deleting autosave...");
+    }
 
     await fs.rmSync(path.join(autosavesFolderPath, name));
   } catch (err) {
-    console.error("Failed to remove autosave.");
-    console.error(err);
+    logger.error({ Name: name, Error: err }, "Failed to delete autosave.");
     res.status(500).json({ error: err });
+  }
+
+  if (serverMaster.successLogs) {
+    logger.info({ Name: name }, "Deleted autosave.");
   }
 
   res.json({ success: true });
@@ -117,16 +174,34 @@ router.delete("/api/removeAutosave", async (req, res) => {
 router.get("/api/getAutosave", async (req, res) => {
   const { name } = req.query;
 
+  if (serverMaster.detailLogs) {
+    logger.info({ Name: name }, "Recieved autosave get request.");
+  }
+
   try {
+    if (serverMaster.detailLogs) {
+      logger.info({ Name: name }, "Getting autosave data...");
+    }
+
     const rawAutosave = await fs.readFileSync(
       path.join(autosavesFolderPath, name),
       "utf-8",
     );
+
+    if (serverMaster.successLogs) {
+      logger.info({ Name: name }, "Got autosave.");
+    }
+    if (serverMaster.detailLogs) {
+      logger.info({ Name: name }, "Sent autosave to Client.");
+    }
+
     res.json(JSON.parse(rawAutosave));
   } catch (err) {
     if (err.code === "ENOENT") {
+      logger.error({ Name: name, Error: err }, "Failed to find autosave.");
       res.status(404).json({ error: "Couldn't find autosave." });
     } else {
+      logger.error({ Name: name, Error: err }, "Failed to get autosave.");
       res.status(500).json({ error: err });
     }
   }